Make Redis start automatically and adjust 9p mount
We don't want to run experiments by hand, make redis start at boot. We
will be using qemu-guest (from the Unikraft paper) to start the guests
so we cannot rely on naming the 9p share 'share', make the mount file
use the script default of 'fs0' instead.

We will be configuring the network interface from the kernel command
line so remove the network file.

Signed-off-by: Eric B Munson <munsoner@bu.edu>